move redirectToPost back to yesod-core
just needed to compile hamlet down to blaze
